pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aluvia-sdk"
version = "1.0.0"
description = "Aluvia SDK for Python - local smart proxy for automation workloads and AI agents"
authors = [{name = "Aluvia", email = "support@aluvia.io"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
keywords = ["proxy", "ai-agent", "playwright", "selenium", "aluvia", "web-scraping"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.24.0",
    "aiohttp>=3.8.0",
    "proxy.py>=2.4.0",
]

[project.optional-dependencies]
playwright = ["playwright>=1.40.0"]
selenium = ["selenium>=4.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "respx>=0.20.0",
]

[project.urls]
Homepage = "https://aluvia.io"
Documentation = "https://docs.aluvia.io"
Repository = "https://github.com/aluvia-connect/sdk-python"
Issues = "https://github.com/aluvia-connect/sdk-python/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["aluvia_sdk*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
